Copy `pakcsrc.default`
macOS's `fs_usage` proves to be an invaluable debugging tool again.

Executing `pakcs` in `sudo -u nobody bash` in a `nix-shell` with `HOME`
set to a temporary directory causes it to crash without an error
message. PAKCS packages itself as a binary SWI-Prolog "saved state",
which makes it hard to inspect. Monitoring `sudo fs_usage | grep swipl`,
however, gives us a clue as to which file could not be found.